More functions, sites-available overwrite choice
toggleSite function for enabling or disabling the site. If the user
wishes to add a new site conf file of the same name as one already in
sites-available, they are asked if they'd like to overwrite it. The
script ends if they choose not to. I regret GitHub isn't properly
showing the differences in the files. Not sure why that's happening.